Make the engine a separate Runnable object in order to provide more flexibility in the way the engine is started.
