+++ trials-engine/src/main/java/org/gardler/biglittlechallenge/trials/EngineControllerApp.java
@@ -1,50 +1,30 @@
 package org.gardler.biglittlechallenge.trials;
 
-import java.net.URI;
-
-import org.gardler.biglittlechallenge.trials.engine.Tournament;
-import org.gardler.biglittlechallenge.trials.engine.TournamentAPI;
-import org.glassfish.grizzly.http.server.HttpServer;
-import org.glassfish.jersey.grizzly2.httpserver.GrizzlyHttpServerFactory;
-import org.glassfish.jersey.server.ResourceConfig;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * The EngineControllerApp is responsible fr starting and running Tournament API instances.
+ * A simple wrapper application that will start an instance of the game Engine.
+ * 
+ * Configuration is via Environment variables:
+ * 
+ * MIN_NUMBER_OF_PLAYERS is the minimum number of players the game should have
  *
  */
 public class EngineControllerApp 
-{
-	
+{	
 	private static Logger logger = LoggerFactory.getLogger(EngineControllerApp.class);
-	
-	// Base URI the Grizzly HTTP server will listen on
-    public static final String BASE_URI = "http://0.0.0.0:8080/api/v0.1/";
-
-        
+	    
     public static void main( String[] args )
     {
         logger.debug("Starting the trials card game.");
-        
         String numString = System.getenv("MIN_NUMBER_OF_PLAYERS");
         if (numString == null) {
         	numString = "2";
         }
-        int players = Integer.parseInt(numString);
-        
-    	logger.debug("Create a tournament server for " + players + " players");
-    	startServer(players);
-    	
-    	while (true) {
-    		
-    	}
+        int players = Integer.parseInt(numString);        
+
+        Engine engine = new Engine(players);
+        engine.run();
     }
-
-	public static HttpServer startServer(int players) {
-		TournamentAPI api = new TournamentAPI(new Tournament(players));
-		final ResourceConfig rc = new ResourceConfig().register(api);
-		return GrizzlyHttpServerFactory.createHttpServer(URI.create(BASE_URI), rc);
-	}
-
 }
